requirements.txt
Pillow==9.4.0
numpy==1.24.1
pip~=22.3.1
requests==2.28.1
setuptools~=65.6.3
torch==1.13.1
torchvision==0.14.1
